+++ truck-meshalgo/src/analyzers/mod.rs
@@ -5,6 +5,7 @@ mod in_out_judge;
 mod point_cloud;
 mod splitting;
 mod topology;
+mod volume;
 
 pub use collision::Collision;
 pub use in_out_judge::IncludingPointInDomain;
@@ -13,3 +14,4 @@ pub use splitting::ExperimentalSplitters;
 pub use splitting::Splitting;
 pub use topology::Topology;
 pub use truck_topology::shell::ShellCondition;
+pub use volume::CalcVolume;