+++ truck-polymesh/src/lib.rs
@@ -19,7 +19,7 @@ use serde::{Deserialize, Serialize};
 /// re-export `truck_base`.
 pub mod base {
     pub use truck_base::{
-        bounding_box::BoundingBox, cgmath64::*, hash, hash::HashGen, tolerance::*,
+        assert_near, bounding_box::BoundingBox, cgmath64::*, hash, hash::HashGen, tolerance::*,
     };
     pub use truck_geotrait::*;
 }